드로잉 플레이에 모델 폴더를 추가합니다. (#107)
* refactor: add models(#106)

* refactor: import models(#106)
