+++ strawberry/src/pages/drawingPlay/models/LineStyle.tsx
@@ -1,0 +1,8 @@
+export type LineStyle = {
+  stroke: string;
+  lineWidth: number;
+  lineStyle: string;
+  roughness: number;
+  bowing: number;
+  strokeWidth: number;
+};
