+++ strawberry/src/pages/drawingPlay/models/Point.tsx
@@ -1,0 +1,1 @@
+export type Point = { x: number; y: number };
